Use ansible module for adding known_hosts
When we combine the use of this role with other additions to
ssh_known_hosts, the existing logic would clobber the global settings.
This would prevent us from registering sane keys for github.com, etc.

To get around this, we'll use a smarter approach to scanning the host
keys and adding to known hosts. The use of the ansible module for this
allows us to update the existing file without clobbering it.